// src/app/user/setting/[settingId]/_components/contentCategory/_components/dnd-kit/Tree/types.ts
import type {MutableRefObject} from 'react';
import type {UniqueIdentifier} from '@dnd-kit/core';

export interface TreeItem {
  id: UniqueIdentifier;
  name: string;
  pos: number;
  count: number;
  collapsed?: boolean;
  isEditing?: boolean;
  children: TreeItem[];
}

export type TreeItems = TreeItem[];

export interface FlattenedItem extends TreeItem {
  parentId: UniqueIdentifier | null;
  depth: number;
  index: number;
}

export type SensorContext = MutableRefObject<{
  items: FlattenedItem[];
  offset: number;
}>;